+++ substrafl/remote/register/register.py
@@ -325,35 +325,35 @@ def register_algo(
         return key
 
 
-def _check_score_function(score_function: typing.Callable):
+def _check_metric_function(metric_function: typing.Callable):
     """Function to check the type and the signature of a given score function.
 
     Args:
-        score_function (typing.Callable): function to check.
+        metric_function (typing.Callable): function to check.
 
     Raises:
-        exceptions.ScoreFunctionTypeError: score_function must be of type "function"
-        exceptions.ScoreFunctionSignatureError: score_function must ONLY contains
+        exceptions.MetricFunctionTypeError: metric_function must be of type "function"
+        exceptions.MetricFunctionSignatureError: metric_function must ONLY contains
             datasamples and predictions_path as parameters
     """
 
-    if not inspect.isfunction(score_function):
-        raise exceptions.ScoreFunctionTypeError("The score_function() must be of type function.")
+    if not inspect.isfunction(metric_function):
+        raise exceptions.MetricFunctionTypeError("The metric_function() must be of type function.")
 
-    signature = inspect.signature(score_function)
+    signature = inspect.signature(metric_function)
     parameters = signature.parameters
 
     if "datasamples" not in parameters:
-        raise exceptions.ScoreFunctionSignatureError(
-            "The score_function() function must contain datasamples as parameter."
+        raise exceptions.MetricFunctionSignatureError(
+            "The metric_function() function must contain datasamples as parameter."
         )
     elif "predictions_path" not in parameters:
-        raise exceptions.ScoreFunctionSignatureError(
-            "The score_function() function must contain predictions_path as parameter."
+        raise exceptions.MetricFunctionSignatureError(
+            "The metric_function() function must contain predictions_path as parameter."
         )
     elif len(parameters) != 2:
-        raise exceptions.ScoreFunctionSignatureError(
-            """The score_function() function must ONLY contains datasamples and predictions_path as
+        raise exceptions.MetricFunctionSignatureError(
+            """The metric_function() function must ONLY contains datasamples and predictions_path as
             parameters."""
         )
 
@@ -362,7 +362,8 @@ def add_metric(
     client: substra.Client,
     permissions: substra.sdk.schemas.Permissions,
     dependencies: Dependency,
-    score_function: typing.Callable,
+    metric_function: typing.Callable,
+    metric_name: typing.Optional[str] = None,
 ) -> str:
     """This function add a metric to the Substra platform using the given score function as the
     algorithm to execute.
@@ -373,18 +374,20 @@ def add_metric(
         client (substra.Client): The substra client.
         permissions (substra.sdk.schemas.Permissions): Permissions for the score function.
         dependencies (Dependency): Score function dependencies.
-        score_function (typing.Callable): function to compute the score from the datasamples and the predictions.
+        metric_function (typing.Callable): function to compute the score from the datasamples and the predictions.
             This function is registered in substra as a metric.
+        metric_name (str, Optional): Optional name chosen by the user to identify the metric. If None,
+            the metric name is set to the 'metric_{metric_function.__name__}'.
 
     Returns:
         str: The metric key of the metric created from the score function.
     """
 
-    _check_score_function(score_function=score_function)
+    _check_metric_function(metric_function=metric_function)
 
     class Metric:
         def score(self, datasamples, predictions_path):
-            return score_function(datasamples=datasamples, predictions_path=predictions_path)
+            return metric_function(datasamples=datasamples, predictions_path=predictions_path)
 
     inputs_metrics = [
         substra.sdk.schemas.AlgoInputSpec(
@@ -422,7 +425,7 @@ def add_metric(
         remote_cls=RemoteDataMethod,
         method_name="score",
         method_parameters={},
-        algo_name="metric_" + score_function.__name__,
+        algo_name=metric_name or "metric_" + metric_function.__name__,
     )
 
     key = register_algo(